+++ lib/pages/inspection/widgets/inspection_card.dart
@@ -1,5 +1,6 @@
 import 'package:cloud/models/inspection/inspection.dart';
 import 'package:cloud/pages/inspection/widgets/collaboration_dialog.dart';
+import 'package:cloud/pages/widgets/confirm_dialog.dart';
 import 'package:cloud/providers/app_provider.dart';
 import 'package:cloud/services/inspection.dart';
 import 'package:flutter/material.dart';
@@ -84,8 +85,14 @@ class InspectionCard extends HookConsumerWidget {
                   const SizedBox(width: 12),
                   InkWell(
                     onTap: () async {
-                      await deleteInspection(inspection.id!);
-                      EasyLoading.showSuccess('删除成功');
+                      final bool isConfirmed = await ConfirmDialog.show(
+                        context,
+                        content: '确定要删除验货任务${inspection.name}？',
+                      );
+                      if (isConfirmed) {
+                        await deleteInspection(inspection.id!);
+                        EasyLoading.showSuccess('删除成功');
+                      }
                     },
                     child: Icon(Icons.delete_outline,
                         size: 18, color: Colors.red[300]),
